chore: Reorganized new data files
