+++ src/data/countChildrenVars.js
